Added canonical name column to the ingredients table
The canonical name of an ingredient is meant to represent the ingredient by name
no matter how the spelling is executed with respect to letter casing. Namely
the ingredients with name `chicken breast` and `Chicken Breast` should be represented
by the same underlying ingredient.

Further by introducing the canonical name for an ingredient we can now avoid duplicate
entries in the database, while still returning the ingredients back to the user in exactly
the same format in which they have saved them into the database. Further it is possible
to change and adapt the spelling in the ingredient name without affecting how it is truly
called.