Add a class to represent the name of the ingredient
As we have introduced a new type to represent the names which we use
in our code base in general it has became useful to create another class
which is going to represent the specific usecase of the CanonicalName class.

This new class is going to prevent us from comparing the names of ingredients
with names of recipes or with names of other named domain concepts.